% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_base_url.R
\name{lp_base_url}
\alias{lp_base_url}
\title{Leanpub Base URL}
\usage{
lp_base_url(secure = TRUE)
}
\arguments{
\item{secure}{Should https be used (may be necessary)}
}
\value{
Character vector (length 1) of URL
}
\description{
Leanpub Base URL
}
\examples{
lp_base_url()
}
