% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_download.R
\name{lp_download}
\alias{lp_download}
\title{Leanpub Download Latest Version}
\usage{
lp_download(
  slug,
  format = c("pdf", "epub"),
  api_key = NULL,
  secure = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{slug}{slug of the project}

\item{format}{format of the book to download}

\item{api_key}{API key for Leanpub, passed to \code{\link{lp_api_key}}}

\item{secure}{passed to \code{\link{lp_base_url}} for https}

\item{verbose}{print diagnostic messages}

\item{...}{additional options to pass to \code{\link[httr]{GET}}}
}
\value{
List of the result of the \code{\link[httr]{GET}} call and
the content
}
\description{
Leanpub Download Latest Version
}
\note{
See \url{https://leanpub.com/help/api}
}
\examples{
if (lp_have_api_key()) {
slug = "biostatmethods"
res = lp_download(slug, verbose = FALSE, nonstop = TRUE, error = FALSE,
format = "epub")
file.exists(res)
}
}
