% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_job_status.R
\name{lp_job_status}
\alias{lp_job_status}
\title{Leanpub Job Status}
\usage{
lp_job_status(slug, api_key = NULL, secure = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{slug}{slug of the project}

\item{api_key}{API key for Leanpub, passed to \code{\link{lp_api_key}}}

\item{secure}{passed to \code{\link{lp_base_url}} for https}

\item{verbose}{print diagnostic messages}

\item{...}{additional options to pass to \code{\link[httr]{GET}}}
}
\value{
List of the result of the \code{\link[httr]{GET}} call and
the content
}
\description{
Leanpub Job Status
}
\note{
See \url{https://leanpub.com/help/api}
}
\examples{
if (lp_have_api_key()) {
stat = lp_job_status(slug = "neuroimagingforstatisticians",
nonstop = TRUE, error = FALSE)
}
}
