% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_links.R
\name{lp_links}
\alias{lp_links}
\title{Leanpub Links}
\usage{
lp_links(slug, api_key = NULL, secure = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{slug}{slug of the project}

\item{api_key}{API key for Leanpub, passed to \code{\link{lp_api_key}}}

\item{secure}{passed to \code{\link{lp_base_url}} for https}

\item{verbose}{print diagnostic messages}

\item{...}{additional options to pass to \code{\link[httr]{GET}}}
}
\value{
List of URLs
}
\description{
Leanpub Links
}
\note{
See \url{https://leanpub.com/help/api}
}
\examples{
if (lp_have_api_key()) {
slug = "biostatmethods"
res = lp_links(slug, nonstop = TRUE, error = FALSE)
}
}
