% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_sales.R
\name{lp_sales}
\alias{lp_sales}
\alias{lp_all_sales}
\title{Leanpub Sales Information}
\usage{
lp_sales(slug, api_key = NULL, secure = TRUE, verbose = TRUE, ...)

lp_all_sales(slug, api_key = NULL, secure = TRUE, verbose = TRUE,
  ...)
}
\arguments{
\item{slug}{slug of the project}

\item{api_key}{API key for Leanpub, passed to \code{\link{lp_api_key}}}

\item{secure}{passed to \code{\link{lp_base_url}} for https}

\item{verbose}{print diagnostic messages}

\item{...}{additional options to pass to \code{\link{GET}}}
}
\value{
List of the result of the \code{\link{GET}} call and
the content
}
\description{
Leanpub Sales Information
}
\note{
See \url{https://leanpub.com/help/api}
}
\examples{
\donttest{
if (lp_have_api_key()) {
lp_all_sales(slug = "biostatmethods", query = list(page=2))
}
}
}
